refactor: move Aspects to another file
